LayoutTests/imported/w3c/web-platform-tests/css/css-transitions/transition-property-002-expected.txt


FAIL parse 'none, all' assert_equals: expected "none, all" but got "all"
FAIL parse 'all, none' assert_equals: expected "all, none" but got "all"
PASS parse 'foobar'
FAIL parse 'all, foobar' assert_equals: expected "all, foobar" but got "foobar"
FAIL parse 'foobar, all' assert_equals: expected "foobar, all" but got "all"
PASS parse 'inherit'
PASS parse 'initial'
